requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
email-validator==2.1.0.post1
gunicorn==21.2.0
google-generativeai
python-dotenv
